Function zcash_note_encryption::batch::try_compact_note_decryption

pub fn try_compact_note_decryption<D: BatchDomain, Output: ShieldedOutput<D, COMPACT_NOTE_SIZE>>(
    ivks: &[D::IncomingViewingKey],
    outputs: &[(D, Output)]
) -> Vec<Option<((D::Note, D::Recipient), usize)>>

Available on crate feature alloc only.

Trial decryption of a batch of notes for light clients with a set of recipients.

This is the batched version of crate::try_compact_note_decryption.

Returns a vector containing the decrypted result for each output, with the same length and in the same order as the outputs were provided, along with the index in the ivks slice associated with the IVK that successfully decrypted the output.