Function zcash_note_encryption::try_output_recovery_with_ovk

pub fn try_output_recovery_with_ovk<D: Domain, Output: ShieldedOutput<D>>(
    domain: &D, 
    ovk: &D::OutgoingViewingKey, 
    output: &Output, 
    cv: &D::ValueCommitment, 
    out_ciphertext: &[u8]
) -> Option<(D::Note, D::Recipient, D::Memo)>

Recovery of the full note plaintext by the sender.

Attempts to decrypt and validate the given enc_ciphertext using the given ovk. If successful, the corresponding note and memo are returned, along with the address to which the note was sent.

Implements Zcash Protocol Specification section 4.19.3.