Struct zcash_note_encryption::NoteEncryption

pub struct NoteEncryption<D: Domain> { /* private fields */ }

A struct containing context required for encrypting Sapling and Orchard notes.

This struct provides a safe API for encrypting Sapling and Orchard notes. In particular, it enforces that fresh ephemeral keys are used for every note, and that the ciphertexts are consistent with each other.

Implements section 4.19 of the Zcash Protocol Specification NB: the example code is only covering the post-Canopy case.

Examples

extern crate ff;
extern crate rand_core;
extern crate zcash_primitives;

use ff::Field;
use rand_core::OsRng;
use zcash_primitives::{
    keys::{OutgoingViewingKey, prf_expand},
    consensus::{TEST_NETWORK, TestNetwork, NetworkUpgrade, Parameters},
    memo::MemoBytes,
    sapling::{
        note_encryption::sapling_note_encryption,
        util::generate_random_rseed,
        Diversifier, PaymentAddress, Rseed, ValueCommitment
    },
};

let mut rng = OsRng;

let diversifier = Diversifier([0; 11]);
let pk_d = diversifier.g_d().unwrap();
let to = PaymentAddress::from_parts(diversifier, pk_d).unwrap();
let ovk = Some(OutgoingViewingKey([0; 32]));

let value = 1000;
let rcv = jubjub::Fr::random(&mut rng);
let cv = ValueCommitment {
    value,
    randomness: rcv.clone(),
};
let height = TEST_NETWORK.activation_height(NetworkUpgrade::Canopy).unwrap();
let rseed = generate_random_rseed(&TEST_NETWORK, height, &mut rng);
let note = to.create_note(value, rseed).unwrap();
let cmu = note.cmu();

let mut enc = sapling_note_encryption::<_, TestNetwork>(ovk, note, to, MemoBytes::empty(), &mut rng);
let encCiphertext = enc.encrypt_note_plaintext();
let outCiphertext = enc.encrypt_outgoing_plaintext(&cv.commitment().into(), &cmu, &mut rng);

Implementations

impl<D: Domain> NoteEncryption<D>

pub fn new(
ovk: Option<D::OutgoingViewingKey>,
note: D::Note,
to: D::Recipient,
memo: D::Memo
) -> Self

Construct a new note encryption context for the specified note, recipient, and memo.

pub fn new_with_esk(
esk: D::EphemeralSecretKey,
ovk: Option<D::OutgoingViewingKey>,
note: D::Note,
to: D::Recipient,
memo: D::Memo
) -> Self

Available on crate feature pre-zip-212 only.

For use only with Sapling. This method is preserved in order that test code be able to generate pre-ZIP-212 ciphertexts so that tests can continue to cover pre-ZIP-212 transaction decryption.

pub fn esk(&self) -> &D::EphemeralSecretKey

Exposes the ephemeral secret key being used to encrypt this note.

pub fn epk(&self) -> &D::EphemeralPublicKey

Exposes the encoding of the ephemeral public key being used to encrypt this note.

pub fn encrypt_note_plaintext(&self) -> [u8; 580]

Generates encCiphertext for this note.

pub fn encrypt_outgoing_plaintext<R: RngCore>(
&self,
cv: &D::ValueCommitment,
cmstar: &D::ExtractedCommitment,
rng: &mut R
) -> [u8; 80]

Generates outCiphertext for this note.