Trait zcash_note_encryption::BatchDomain

pub trait BatchDomain: Domain {
    fn batch_kdf<'a>(
        items: impl Iterator<Item = (Option<Self::SharedSecret>, &'a EphemeralKeyBytes)>
    ) -> Vec<Option<Self::SymmetricKey>> { ... }

    fn batch_epk(
        ephemeral_keys: impl Iterator<Item = EphemeralKeyBytes>
    ) -> Vec<(Option<Self::EphemeralPublicKey>, EphemeralKeyBytes)> { ... }
}

Provided methods

fn batch_kdf<'a>(
items: impl Iterator<Item = (Option<Self::SharedSecret>, &'a EphemeralKeyBytes)>
) -> Vec<Option<Self::SymmetricKey>>

Computes Self::kdf on a batch of items.

For each item in the batch, if the shared secret is None, this returns None at that position.

fn batch_epk(
ephemeral_keys: impl Iterator<Item = EphemeralKeyBytes>
) -> Vec<(Option<Self::EphemeralPublicKey>, EphemeralKeyBytes)>

Computes Self::epk on a batch of ephemeral keys.

This is useful for protocols where the underlying curve requires an inversion to parse an encoded point.

For usability, this returns tuples of the ephemeral keys and the result of parsing them.