Crate zcash_note_encryption

Implementation of in-band secret distribution abstractions for Zcash transactions. The implementations here provide functionality that is shared between the Sapling and Orchard protocols.

Modules

batch

APIs for batch trial decryption.

Structs

EphemeralKeyBytes

NoteEncryption

A struct containing context required for encrypting Sapling and Orchard notes.

NotePlaintextBytes

OutPlaintextBytes

OutgoingCipherKey

A symmetric key that can be used to recover a single Sapling or Orchard output.

Enums

NoteValidity

Constants

AEAD_TAG_SIZE

COMPACT_NOTE_SIZE

ENC_CIPHERTEXT_SIZE

NOTE_PLAINTEXT_SIZE

OUT_CIPHERTEXT_SIZE

OUT_PLAINTEXT_SIZE

Traits

Domain

ShieldedOutput

Functions

try_compact_note_decryption

Trial decryption of the compact note plaintext by the recipient for light clients.

try_note_decryption

Trial decryption of the full note plaintext by the recipient.

try_output_recovery_with_ock

Recovery of the full note plaintext by the sender.

try_output_recovery_with_ovk

Recovery of the full note plaintext by the sender.