Trait zcash_note_encryption::Domain

pub trait Domain {

    type EphemeralSecretKey: ConstantTimeEq;
    type EphemeralPublicKey;
    type SharedSecret;
    type SymmetricKey: AsRef<[u8]>;
    type Note;
    type Recipient;
    type DiversifiedTransmissionKey;
    type IncomingViewingKey;
    type OutgoingViewingKey;
    type ValueCommitment;
    type ExtractedCommitment;
    type ExtractedCommitmentBytes: Eq + for<'a> From<&'a Self::ExtractedCommitment>;
    type Memo;
    fn derive_esk(note: &Self::Note) -> Option<Self::EphemeralSecretKey>;

    fn get_pk_d(note: &Self::Note) -> Self::DiversifiedTransmissionKey;

    fn ka_derive_public(
        note: &Self::Note, 
        esk: &Self::EphemeralSecretKey
    ) -> Self::EphemeralPublicKey;

    fn ka_agree_enc(
        esk: &Self::EphemeralSecretKey, 
        pk_d: &Self::DiversifiedTransmissionKey
    ) -> Self::SharedSecret;

    fn ka_agree_dec(
        ivk: &Self::IncomingViewingKey, 
        epk: &Self::EphemeralPublicKey
    ) -> Self::SharedSecret;

    fn kdf(
        secret: Self::SharedSecret, 
        ephemeral_key: &EphemeralKeyBytes
    ) -> Self::SymmetricKey;

    fn note_plaintext_bytes(
        note: &Self::Note, 
        recipient: &Self::Recipient, 
        memo: &Self::Memo
    ) -> NotePlaintextBytes;

    fn derive_ock(
        ovk: &Self::OutgoingViewingKey, 
        cv: &Self::ValueCommitment, 
        cmstar_bytes: &Self::ExtractedCommitmentBytes, 
        ephemeral_key: &EphemeralKeyBytes
    ) -> OutgoingCipherKey;

    fn outgoing_plaintext_bytes(
        note: &Self::Note, 
        esk: &Self::EphemeralSecretKey
    ) -> OutPlaintextBytes;

    fn epk_bytes(epk: &Self::EphemeralPublicKey) -> EphemeralKeyBytes;

    fn epk(
        ephemeral_key: &EphemeralKeyBytes
    ) -> Option<Self::EphemeralPublicKey>;

    fn check_epk_bytes<F: Fn(&Self::EphemeralSecretKey) -> NoteValidity>(
        note: &Self::Note, 
        check: F
    ) -> NoteValidity;

    fn cmstar(note: &Self::Note) -> Self::ExtractedCommitment;

    fn parse_note_plaintext_without_memo_ivk(
        &self, 
        ivk: &Self::IncomingViewingKey, 
        plaintext: &[u8]
    ) -> Option<(Self::Note, Self::Recipient)>;

    fn parse_note_plaintext_without_memo_ovk(
        &self, 
        pk_d: &Self::DiversifiedTransmissionKey, 
        esk: &Self::EphemeralSecretKey, 
        ephemeral_key: &EphemeralKeyBytes, 
        plaintext: &[u8]
    ) -> Option<(Self::Note, Self::Recipient)>;

    fn extract_memo(&self, plaintext: &[u8]) -> Self::Memo;

    fn extract_pk_d(
        out_plaintext: &[u8; 64]
    ) -> Option<Self::DiversifiedTransmissionKey>;

    fn extract_esk(out_plaintext: &[u8; 64]) -> Option<Self::EphemeralSecretKey>;

    fn batch_kdf<'a>(
        items: impl Iterator<Item = (Option<Self::SharedSecret>, &'a EphemeralKeyBytes)>
    ) -> Vec<Option<Self::SymmetricKey>> { ... }

    fn batch_epk(
        ephemeral_keys: impl Iterator<Item = EphemeralKeyBytes>
    ) -> Vec<(Option<Self::EphemeralPublicKey>, EphemeralKeyBytes)> { ... }

}

Associated Types

type EphemeralSecretKey: ConstantTimeEq

type EphemeralPublicKey

type SharedSecret

type SymmetricKey: AsRef<[u8]>

type Note

type Recipient

type DiversifiedTransmissionKey

type IncomingViewingKey

type OutgoingViewingKey

type ValueCommitment

type ExtractedCommitment

type ExtractedCommitmentBytes: Eq + for<'a> From<&'a Self::ExtractedCommitment>

type Memo

Required methods

fn derive_esk(note: &Self::Note) -> Option<Self::EphemeralSecretKey>

fn get_pk_d(note: &Self::Note) -> Self::DiversifiedTransmissionKey

fn ka_derive_public(
note: &Self::Note,
esk: &Self::EphemeralSecretKey
) -> Self::EphemeralPublicKey

fn ka_agree_enc(
esk: &Self::EphemeralSecretKey,
pk_d: &Self::DiversifiedTransmissionKey
) -> Self::SharedSecret

fn ka_agree_dec(
ivk: &Self::IncomingViewingKey,
epk: &Self::EphemeralPublicKey
) -> Self::SharedSecret

fn kdf(
secret: Self::SharedSecret,
ephemeral_key: &EphemeralKeyBytes
) -> Self::SymmetricKey

fn note_plaintext_bytes(
note: &Self::Note,
recipient: &Self::Recipient,
memo: &Self::Memo
) -> NotePlaintextBytes

fn derive_ock(
ovk: &Self::OutgoingViewingKey,
cv: &Self::ValueCommitment,
cmstar_bytes: &Self::ExtractedCommitmentBytes,
ephemeral_key: &EphemeralKeyBytes
) -> OutgoingCipherKey

fn outgoing_plaintext_bytes(
note: &Self::Note,
esk: &Self::EphemeralSecretKey
) -> OutPlaintextBytes

fn epk_bytes(epk: &Self::EphemeralPublicKey) -> EphemeralKeyBytes

fn epk(ephemeral_key: &EphemeralKeyBytes) -> Option<Self::EphemeralPublicKey>

fn check_epk_bytes<F: Fn(&Self::EphemeralSecretKey) -> NoteValidity>(
note: &Self::Note,
check: F
) -> NoteValidity

fn cmstar(note: &Self::Note) -> Self::ExtractedCommitment

fn parse_note_plaintext_without_memo_ivk(
&self,
ivk: &Self::IncomingViewingKey,
plaintext: &[u8]
) -> Option<(Self::Note, Self::Recipient)>

fn parse_note_plaintext_without_memo_ovk(
&self,
pk_d: &Self::DiversifiedTransmissionKey,
esk: &Self::EphemeralSecretKey,
ephemeral_key: &EphemeralKeyBytes,
plaintext: &[u8]
) -> Option<(Self::Note, Self::Recipient)>

fn extract_memo(&self, plaintext: &[u8]) -> Self::Memo

fn extract_pk_d(
out_plaintext: &[u8; 64]
) -> Option<Self::DiversifiedTransmissionKey>

fn extract_esk(out_plaintext: &[u8; 64]) -> Option<Self::EphemeralSecretKey>

Provided methods

fn batch_kdf<'a>(
items: impl Iterator<Item = (Option<Self::SharedSecret>, &'a EphemeralKeyBytes)>
) -> Vec<Option<Self::SymmetricKey>>

Computes Self::kdf on a batch of items.

For each item in the batch, if the shared secret is None, this returns None at that position.

fn batch_epk(
ephemeral_keys: impl Iterator<Item = EphemeralKeyBytes>
) -> Vec<(Option<Self::EphemeralPublicKey>, EphemeralKeyBytes)>

Computes Self::epk on a batch of ephemeral keys.

This is useful for protocols where the underlying curve requires an inversion to parse an encoded point.

For usability, this returns tuples of the ephemeral keys and the result of parsing them.