Trait zcash_note_encryption::ShieldedOutput

pub trait ShieldedOutput<D: Domain, const CIPHERTEXT_SIZE: usize> {
    fn ephemeral_key(&self) -> EphemeralKeyBytes;
    fn cmstar_bytes(&self) -> D::ExtractedCommitmentBytes;
    fn enc_ciphertext(&self) -> &[u8; CIPHERTEXT_SIZE];
}

Trait that provides access to the components of an encrypted transaction output.

Implementations of this trait are required to define the length of their ciphertext field. In order to use the trial decryption APIs in this crate, the length must be either ENC_CIPHERTEXT_SIZE or COMPACT_NOTE_SIZE.

Required Methods

fn ephemeral_key(&self) -> EphemeralKeyBytes

Exposes the ephemeral_key field of the output.

fn cmstar_bytes(&self) -> D::ExtractedCommitmentBytes

Exposes the cmu_bytes or cmx_bytes field of the output.

fn enc_ciphertext(&self) -> &[u8; CIPHERTEXT_SIZE]

Exposes the note ciphertext of the output.